+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Supported the URL register and navigation.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initializing project.